% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{build_K}
\alias{build_K}
\title{Make Gram Matrix}
\usage{
build_K(x, y, k, theta)
}
\arguments{
\item{x}{input matrix 1}

\item{y}{input matrix 2}

\item{k}{R function to apply covariance function for elements of \code{x} and \code{y}}

\item{theta}{hyperparameter vector to input to \code{k}}
}
\description{
Compute the Gram (covariance) matrix
}
\details{
The kernel function \code{k} needs to have arguments corresponding to  \code{x}, \code{y}, \code{theta} in that order.

\code{build_K} computes the gram matrix for a given \code{k}. For a faster, parallel approach,
\code{make_gram_par} uses a pre-determined Gaussian covariance function.
}
