% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ell_ss_sample}
\alias{ell_ss_sample}
\title{Elliptical Slice Sampling}
\usage{
ell_ss_sample(y, f, K)
}
\arguments{
\item{y}{binary output vector in -1, +1}

\item{f}{latent variable for Gaussian process}

\item{K}{Gram matrix}
}
\value{
new proposal for \eqn{f}
}
\description{
Elliptical method to sample latent variables \code{f}
}
\details{
Sampling of latent variables \code{f} constrained to an ellipse centred at the current mean
\eqn{f}, with eccentricity of the ellipse determined by the covariance matrix \eqn{\Sigma}.
}
