% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geometric_mean}
\alias{geometric_mean}
\title{Geometric mean}
\usage{
geometric_mean(x, NA.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector of non-negative numbers.}
}
\description{
Geometric mean
}
\note{
\code{prod()} can under or overflow very quickly.
}
