% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel-SE.R
\name{gradSEmat}
\alias{gradSEmat}
\title{Gradient of the squared exponential kernel, matrix version}
\usage{
gradSEmat(x, len = 1, K = NULL)
}
\arguments{
\item{x}{Vector of scalar parameters, or matrix of vector parameters.}

\item{len}{Length-scale of correlation, isotropic (scalar) or separable (vector), defaults to 1.}

\item{K}{Kernel matrix, if already computed}
}
\value{
a list of partial derivatives of the correlation matrix w.r.t. each lengthscale.
}
\description{
Gradient of the squared exponential kernel, matrix version
}
