% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivm.R
\name{ivm_subset_selection}
\alias{ivm_subset_selection}
\title{IVM Subset Selection}
\usage{
ivm_subset_selection(X, kernel, theta, nsub, noise = 0.001)
}
\arguments{
\item{X}{predictor matrix}

\item{kernel}{kernel function for the covariance in \code{X}}

\item{theta}{hyperparameter vector}

\item{nsub}{desired subset size}

\item{noise}{Gaussian variance (noise) in observations}
}
\value{
a vector of indices of \code{y} giving the locations for the subset
}
\description{
Implementation of IVM subset selection for Gaussian process classification
}
\details{
This performs the 'Information Vector Machine' (IVM) algorithm to choose a
subset of data to reduce computation time when fitting a Gaussian Process Classification
model.
}
