% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{laplace_approx}
\alias{laplace_approx}
\title{Laplace Approximation}
\usage{
laplace_approx(y, K)
}
\arguments{
\item{y}{binary output vector in -1, +1}

\item{K}{Gram matrix}
}
\value{
list containing two elements:
\item{\code{f_hat}}{mean of the approximate Gaussian distribution}
\item{\code{sigma_hat}}{covariance matrix of the approximate Gaussian distribution}
}
\description{
Laplace approximation of the posterior of the latent Variables \eqn{f}.
}
\details{
Obtain a Laplace approximate of the posterior of \deqn{
p(f | y, \theta)
}
by approximating
the distribution with a Gaussian \eqn{N(f | \mu_q, \Sigma_q)}.

This amounts to an iterative procedure, iterating a Newton-Raphson formula:
\deqn{
f_{new} = f - (\nabla_f \nabla_f \Psi(f))^{-1} \nabla_f \Psi(f),
}
where \deqn{
\Psi(f) = log p(y | f) + log p (f | \theta) + const.
}
Beginning from \eqn{f=0}, this
is iterated until convergence. That is, when the distance between \eqn{f} and \eqn{f_{new}} becomes
sufficiently small.
}
