% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_gram_par}
\alias{make_gram_par}
\title{Make Gram Matrix (Gaussian covariance)}
\usage{
make_gram_par(x, y, theta)
}
\arguments{
\item{x}{input matrix 1}

\item{y}{input matrix 2}

\item{theta}{hyperparameter vector}
}
\description{
Compute the Gram (covariance) matrix in parallel with a Gaussian covariance function
}
\details{
This is a faster, parallel approach to build the Gram matrix using a pre-determined Gaussian covariance function.
\code{build_K} computes the gram matrix for a given \code{k}, but is slower and not in parallel.
}
