% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRPC.R
\name{MRPC}
\alias{MRPC}
\title{Mendelian Randomization Power Calculation}
\usage{
MRPC(x, ...)
}
\arguments{
\item{OR}{odds ratio}

\item{rsq}{The variance explained by the additive effect on the genotype}

\item{N}{The total sample number}

\item{Ncase}{The number of case}

\item{Ncontrol}{The number of control}

\item{outcome}{either 'linear' or 'binary'}

\item{K}{the number of instrumental variable}
}
\description{
It calculate the Mendelian randomization power based on the odds ratio, the variance explained by genotype,
and number of case and control.

The power to detect a causal effect can be calculated as :
\itemize{
 \item {For linear} : {\eqn{\Phi(\beta_1\rho_{GX} \sqrt{N} -z_{(1- {\alpha\over 2})})}}
 \item {For binary} : {\eqn{\Phi(\beta_1\rho_{GX} \sqrt{N P(Y=1)P(Y=0)} -z_{(1- {\alpha\over 2})})}}
}
where \eqn{\rho_{GX}} is the variance explained by the genotype (\eqn{\approx 2f(1-f)\beta^2}).
It can be calculated using effect size \eqn{\beta} and minor allele frequency.
}
\examples{
MRPC(OR=c(1.2,1.3,1.4,1.5,1.6), rsq=c(0.1, 0.2, 0.3), N=1000, Ncase=500, Ncontrol=500, outcome='binary', K=c(1,2,3))
MRPC(OR=c(1.2,1.3,1.65,1.87), rsq=c(0.111111111, 0.022), N=1000, Ncase=500, Ncontrol=500, outcome='binary', K=c(1))
}
\seealso{
\link{https://pubmed.ncbi.nlm.nih.gov/24608958/}
}
