% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{plot.gpc}
\alias{plot.gpc}
\title{Plot Gaussian Process Classification Model}
\usage{
\method{plot}{gpc}(x, f = FALSE, ...)
}
\arguments{
\item{x}{object of type '\code{gpc}', output from \code{\link{gpc}}}

\item{...}{further arguments to be passed to plot}
}
\value{
trace and density plot of the model fit
}
\description{
Plot Gaussian Process Classification Model
}
