% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{predict.gpc}
\alias{predict.gpc}
\title{Predict from Gaussian Process Classification Model}
\usage{
\method{predict}{gpc}(object, newdata, ...)
}
\arguments{
\item{object}{object of type '\code{gpc}', output from \code{\link{gpc}}}

\item{newdata}{new data matrix, of same dimension as \code{X} used for fitting}

\item{...}{further arguments to pass to \code{predict}}
}
\value{
a vector of probabilities corresponding to the positive and negative class
}
\description{
Predict from Gaussian Process Classification Model
}
