% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_gp}
\alias{predict_gp}
\title{Rcpp Predict Gaussian Process Classification using MCMC}
\usage{
predict_gp(y, X, newdata, kernel, fit, print_every)
}
\arguments{
\item{y}{binary output vector in {-1, +1}}

\item{X}{predictor matrix}

\item{newdata}{new data to predict from, same dimension as \code{X}}

\item{kernel}{R function taking two inputs to calculate the kernel matrix}

\item{fit}{list containing samples from the mcmc chains for \code{f} and \code{theta}}

\item{print_every}{number of steps until progress is printed to the console}
}
\value{
fitted probabilities corresponding to the positive and negative class for each data point in \code{newdata}
}
\description{
Pseudo marginal approach to fitting a Gaussian process classification model using Markov Chain Monte Carlo (MCMC)
}
