% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{print.gpc}
\alias{print.gpc}
\title{Print Gaussian Process Classification Model}
\usage{
\method{print}{gpc}(x, ...)
}
\arguments{
\item{x}{object of type '\code{gpc}', output from \code{\link{gpc}}}

\item{...}{further arguments to be passed to print}
}
\description{
Print Gaussian Process Classification Model
}
