% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safeNonNeg}
\alias{safeNonNeg}
\title{Safely set negative (eigen-)values to zero.}
\usage{
safeNonNeg(eigs)
}
\arguments{
\item{eigs}{A vector of real eigenvalues in decreasing order.}
}
\description{
Safely set negative (eigen-)values to zero.
}
