% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ToleranceLevel}
\alias{ToleranceLevel}
\title{Readable tolerance level for \code{optim()}.}
\usage{
ToleranceLevel(x)
}
\arguments{
\item{x}{Precision for optimization.}
}
\description{
Readable tolerance level for \code{optim()}.
}
