\name{GPC-package}
\alias{GPC-package}
\alias{GPC}
\docType{package}
\title{
\packageTitle{GPC}
}
\description{
\packageDescription{GPC}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GPC}
\packageIndices{GPC}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{GPC}

Maintainer: \packageMaintainer{GPC}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
