% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spam}
\alias{spam}
\title{Spam dataset}
\format{A data frame with 4601 rows and 58 variables.
\describe{
\item{y}{binary response variable: +1 if spam, -1 if not}
\item{X}{57 attributes detailing word frequencies and character frequencies in each e-mail}
}}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/spambase}
}
\usage{
data(spam)
}
\description{
A dataset containing a collection of spam e-mails, often used in classification examples.
}
\keyword{datasets}
